Add and refactor classes for enhanced metrics tracking and schema validation
- Introduced new classes (`Step`, `Task`, `Issue`, `Project`) with structured metrics aggregation and lazy-loading for enhanced performance.
- Implemented Zod schemas (`JunieStepSchema`, `JunieTaskSchema`, etc.) to validate structured data consistency.
- Refactored `JetBrains` class to use modularized classes and schema validations.
- Updated `test-jetbrains.ts` for metric logging and added memory usage tracking.
